Read configmaps from the pod namespace
This commit limits the used configmaps to namespace in which a pod
is created.

Also, the special label property is removed in favor of using simply
the name from the CNI spec.
Configmaps are filtered based on the value of the cni spec name
property.

For example, if CNI spec name is: br1-with-cidr-filtering,
then the configmap needs to be labelled with:
metadata:
  labels:
    br1-with-cidr-filtering: ""

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>