Fix issue with bridge prerouting being skipped
Due to a feature enabled by default:
"net.bridge.bridge-nf-call-iptables=1"
packets get stolen in bridge prerouting, and then are passed to
iptables (as if we were routing...), and then reinjected into bridge.

call-iptables is still needed for backwards compatibility, but changes packet flow and
impacts performance in negative way. The best practice is to always use priority -1 and
higher to avoid this.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>