Limit API requests to Kube API
This change limits the number of API requests to only one
listing of referenced configMaps - instead of getting the configMaps
one by one.

To simplify working with the configMap, renamed the config.json key
to only config, so that it's not needed to escape the dot in the key
name.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>